// frontend/src/types/generated/BatchForecast.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutoName } from "./AutoName";
import type { Resolution } from "./Resolution";

export type BatchForecast = { name: string; nameEnAuto: AutoName | null; cnStart: number; cnEnd: number; resolution: Resolution };